requirement.txt
asgiref==3.7.2
Django==4.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
inflection==0.5.1
packaging==23.1
Pillow==9.5.0
psycopg2-binary==2.9.6
PyJWT==2.7.0
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
typing_extensions==4.6.3
uritemplate==4.1.1
